Generic database (catalog as owner / MD identifier quote)

+++ src/com/dci/intellij/dbn/database/common/metadata/def/DBDataTypeMetadata.java
@@ -10,7 +10,7 @@ public interface DBDataTypeMetadata {
 
     String getDataTypeProgram() throws SQLException;
 
-    int getDataLength() throws SQLException;
+    long getDataLength() throws SQLException;
 
     int getDataPrecision() throws SQLException;
